Take DomDocument into its own file
