+++ XmlParser.cpp
@@ -1,6 +1,7 @@
 #include "stdafx.h"
 #include "XmlParser.h"
 
+#include "DomDocument.h"
 #include "SystemError.h"
 #include "encoding.h"
 
@@ -9,8 +10,8 @@
 #include <msxml6.h>
 #pragma comment(lib, "msxml6.lib")
 
-namespace
-{
+//namespace
+//{
 
 #define RELEASE(iface)    \
     if (iface)            \
@@ -18,119 +19,16 @@ namespace
         iface->Release(); \
         iface = NULL;     \
     }
-
-    template <class T>
-    struct Destroyer
-    {
-        void operator()(T *iface)
-        {
-            iface->Release();
-        }
-    };
-
-    class XML_Document
-    {
-      public:
-        /** Creates an XML document from the supplied filename */
-        explicit XML_Document(std::wstring const &filename)
-        {
-            create_xml_doc_obj();
-
-            BSTR bstrValue{bstr_t(filename.c_str())};
-            CComVariant value(bstrValue);
-
-            VARIANT_BOOL resultCode = FALSE;
-            HRESULT hr = document_->load(value, &resultCode);
-
-            check_load_results(resultCode, hr);
-        }
-
-        /** Creates an XML document from the supplied UTF8 string */
-        explicit XML_Document(std::string const &xml)
-        {
-            create_xml_doc_obj();
-
-            std::wstring string = Encoding::toUnicode(xml);
-            BSTR bstrValue{(bstr_t(string.c_str()))};
-
-            VARIANT_BOOL resultCode = FALSE;
-            HRESULT hr = document_->loadXML(bstrValue, &resultCode);
-
-            check_load_results(resultCode, hr);
-        }
-
-        ~XML_Document()
-        {
-        }
-
-        //Klunky? Remove it
-        IXMLDOMDocument2 *operator->()
-        {
-            return document_.get();
-        }
-
-      private:
-        void create_xml_doc_obj()
-        {
-            IXMLDOMDocument2 *tmp = nullptr;
-            HRESULT hr = CoCreateInstance(__uuidof(DOMDocument), NULL, CLSCTX_SERVER, IID_IXMLDOMDocument2, (LPVOID *)&tmp);
-            if (!SUCCEEDED(hr))
-            {
-                throw ::Linter::SystemError(hr, "Linter: Can't create IID_IXMLDOMDocument2");
-            }
-            document_.reset(tmp);
-
-            hr = document_->put_async(VARIANT_FALSE);
-            if (!SUCCEEDED(hr))
-            {
-                throw ::Linter::SystemError("Linter: Can't XMLDOMDocument2::put_async");
-            }
-        }
-
-        void check_load_results(VARIANT_BOOL resultcode, HRESULT hr)
-        {
-            if (!SUCCEEDED(hr))
-            {
-                throw ::Linter::SystemError(hr);
-            }
-            if (resultcode != VARIANT_TRUE)
-            {
-                CComPtr<IXMLDOMParseError> error;
-                (void)document_->get_parseError(&error);
-
-                if (error)
-                {
-                    BSTR reason;
-                    error->get_reason(&reason);
-                    long line;
-                    error->get_line(&line);
-                    long column;
-                    error->get_linepos(&column);
-                    char buff[256];
-                    std::snprintf(buff,
-                        sizeof(buff),
-                        "Invalid XML in linter.xml at line %d col %d: %s",
-                        line,
-                        column,
-                        static_cast<std::string>(static_cast<bstr_t>(reason)).c_str());
-                    throw std::runtime_error(buff);
-                }
-            }
-        }
-
-        //Replace this once I have moved selectnodes into here
-        std::unique_ptr<IXMLDOMDocument2, Destroyer<IXMLDOMDocument2>> document_;
-        //CComPtr<IXMLDOMDocument2> document_;
-    };
-
-}    // namespace
+//}    // namespace
 
 std::vector<XmlParser::Error> XmlParser::getErrors(const std::string &xml)
 {
-    XML_Document XMLDocument{xml};
+    ::Linter::DomDocument XMLDocument;
+
+    XMLDocument.load_from_string(xml);
 
     // <error line="12" column="19" severity="error" message="Unexpected identifier" source="jscs" />
-    std::unique_ptr<IXMLDOMNodeList, Destroyer<IXMLDOMNodeList>> XMLNodeList;
+    std::unique_ptr<IXMLDOMNodeList, ::Linter::Destroyer<IXMLDOMNodeList>> XMLNodeList;
     {
         IXMLDOMNodeList *tmp = nullptr;
         HRESULT hr = XMLDocument->selectNodes(bstr_t(L"//error"), &tmp);
@@ -186,7 +84,9 @@ XmlParser::Settings XmlParser::getLinters(std::wstring file)
 
     try
     {
-        XML_Document XMLDocument(file);
+        ::Linter::DomDocument XMLDocument;
+
+        XMLDocument.load_from_file(file);
 
         HRESULT hr = XMLDocument->selectNodes(bstr_t(L"//style"), &styleNode);
         if (!SUCCEEDED(hr))
